(MODULES-5013) Use custom acceptance default node
Previously in commit 6dc1284121 the module was updated to use modsync however
the default node file was changed.  This caused Jenkins CI acceptance test to
fail as it was not setup for a vagrant based beaker job.  This commit
reinstates the original default node file and marks it unmanaged for modsync
so the error will not return.